requirements.txt
deeprob_kit==1.1.0
evaluate==0.2.2
matplotlib==3.5.1
numpy==1.21.5
pandas==1.4.2
pgmpy==0.1.18
scikit_learn==1.1.2
scipy==1.7.3
tqdm==4.62.3
ufal.chu_liu_edmonds==1.0.1
